Added syncnode

+++ UTM.Book/UTM.BookAPI/Controllers/BookController.cs
@@ -1,6 +1,8 @@
 ﻿using UTM.Domain.Models;
 using Microsoft.AspNetCore.Mvc;
 using UTM.Service;
+using UTM.BookAPI.Services;
+using AutoMapper;
 
 namespace UTM.BookAPI.Controllers
 {
@@ -9,9 +11,14 @@ namespace UTM.BookAPI.Controllers
     public class BookController : Controller
     {
         private readonly IBookService _bookService;
-        public BookController(IBookService bookService)
+        private readonly ISyncService<Book> _syncService;
+        private readonly IMapper _mapper;
+
+        public BookController(IBookService bookService, ISyncService<Book> syncService, IMapper mapper)
         {
             _bookService = bookService;
+            _syncService = syncService;
+            _mapper = mapper;
         }
         [HttpGet]
         public async Task<List<Book>> GetAllBooks()
@@ -28,6 +35,8 @@ namespace UTM.BookAPI.Controllers
         public async Task<IActionResult> Create(BookIn bookIn)
         {
             var result = await _bookService.AddBook(bookIn);
+
+            await _syncService.Upsert(result);
             
             return Ok(result);
         }
@@ -35,16 +44,30 @@ namespace UTM.BookAPI.Controllers
         [HttpPut]
         public async Task<IActionResult> Upsert(Book book)
         {
-            await _bookService.UpsertBook(book);
+            var result = await _bookService.UpsertBook(book);
+
+            await _syncService.Upsert(result);
 
             return Ok();
         }
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(Guid id)
         {
+            var bookOut = await _bookService.GetBookById(id);
+
+            if(bookOut == null)
+            {
+                return NotFound();
+            }
+
+            var book = _mapper.Map<Book>(bookOut);
+            book.LastChangedAt = DateTime.UtcNow;
+
             await _bookService.DeleteBook(id);
 
-            return Ok();
+            await _syncService.Delete(book);
+            
+            return Ok("Deleted " + id);
         }
 
         [HttpDelete("all")]
@@ -53,6 +76,36 @@ namespace UTM.BookAPI.Controllers
             await _bookService.DeleteAll();
 
             return Ok();
+        }
+
+        [HttpPut("sync")]
+        public async Task<IActionResult> UpsertSync(Book book)
+        {
+            var existingBook = await _bookService.GetBookById(book.Id);
+
+            if(existingBook == null || existingBook.LastChangedAt > book.LastChangedAt)
+            {
+                await _bookService.UpsertBook(book);
+            }
+            return Ok();
+        }
+
+        [HttpDelete("sync")]
+        public async Task<IActionResult> DeleteSync(Book book)
+        {
+            var existingBook = await _bookService.GetBookById(book.Id);
+
+            if (existingBook != null || existingBook.LastChangedAt > book.LastChangedAt)
+            {
+                await _bookService.DeleteBook(book.Id);
+            }
+            return Ok();
+        }
+
+        [HttpGet("test")]
+        public async Task Test()
+        {
+            await _bookService.Test();
         }
     }
 }
